Rewrote Action Button (Audio Tab)
I completely rewrote the Action button on the rows in the Audio Tab because of YTMusicUltimate had a better implementation compared to what YouTube Reborn originally had. This won’t be on the All & Video tabs yet. As this is an experimental upgrade. Also I added the Modify File Name Functionality in this. 😉